Add a separate example for zero delay touchable
