+++ example/app/src/examples/SortableFlex/tests/index.ts
@@ -1,1 +1,2 @@
 export { default as ComplexLayoutExample } from './ComplexLayoutExample';
+export { default as ZeroDelayActivationTouchableExample } from './ZeroDelayActivationTouchableExample';
